Complete UI for the calculator project
Basic structure in index.html with couple of divs
to create the different secions of the calculator.

Used the style.css to, well, style the calculator.
I went for a theme almost similar to the calculator
you find on apple devices, so it has a modern and
minimalistic feel.

Dynamically created and added the buttons using js
instead of overloading the index.html file with div
and button elements. Used the addButton() function
in addButtons.js.

Imported and the addButtons.js file in the script.js
then called the addButton() function.